Fixed the local switcher for non-translatable entities
